feat: Toggle Group  (#103)
* Toggle Group inIt

* feat: enhance ToggleGroup and ToggleItem components

Add styling and structure to the ToggleGroup component, enabling 
flexbox layout and spacing. Update ToggleItem styles for better 
visual feedback on hover and disabled states. Refactor both 
components to accept additional class names for customization. 
Export components for easier access in other modules.

* feat: update component styles and add ToggleGroup stories

Remove unused styles from index.scss and add ToggleGroup and ToggleItem 
styles. Introduce stories for ToggleGroup to demonstrate its usage 
with single and multiple selection options, including disabled states 
and icon support. This enhances the component library by providing 
clear examples and improving the organization of styles.

* feat(styles): Remove duplicate SCSS imports

The changes remove duplicate imports of the `messageLoading.scss` and `copilotShell.scss` files in the `index.scss` file. This ensures that the styles are only imported once, which can improve the performance and maintainability of the application.

* feat(storybook): add preflight CSS file

This commit adds a new CSS file `preflight.css` to the Storybook configuration. The preflight CSS file provides a consistent baseline for the UI components by resetting the default styles and applying a set of sensible defaults. This ensures that the UI components are rendered consistently across different browsers and environments.

The key changes include:

- Resetting the box-sizing, margins, and padding for all elements.
- Applying a consistent line-height and font-family settings.
- Normalizing the appearance of various HTML elements like headings, links, and lists.
- Addressing known browser inconsistencies and quirks.

By including this preflight CSS file, the UI components in the Storybook can be developed and tested with a solid foundation, making it easier to create a cohesive and visually consistent user interface.