✨ feat(MarkDownRenderer): add margin to crayon-table-container (#114)
* ✨ feat(MarkDownRenderer): add margin to crayon-table-container

Adds a margin-bottom of 20px to the .crayon-table-container class
within the MarkDownRenderer component. This ensures proper spacing
between the table and any subsequent content in the rendered
markdown.

* max-width: none

* ✨ feat(MarkDownRenderer): add optional className prop

Adds an optional `className` prop to the `MarkDownRenderer` component to allow
for additional styling to be applied to the rendered markdown content.
This provides more flexibility for consumers of the component to customize
the appearance of the markdown rendering.

* version bump